chore: add jsconfig for path aliases
