+++ components/ProductCard.jsx
@@ -1,14 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 
-interface Props {
-  title: string;
-  image: string;
-  description: string;
-  price: string;
-}
-
-export default function ProductCard({ title, image, description, price }: Props) {
+export default function ProductCard({ title, image, description, price }) {
   return (
     <div className="rounded border p-4 text-center">
       <Image src={image} alt={title} width={400} height={300} className="mx-auto" />
