+++ next.config.mjs
@@ -1,4 +1,3 @@
-/** @type {import('next').NextConfig} */
 const nextConfig = {};
 
 export default nextConfig;
