+++ package.json
@@ -19,11 +19,8 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
-    "@types/node": "^20.11.30",
-    "@types/react": "^19.0.0",
     "eslint": "^9",
     "eslint-config-next": "15.4.6",
-    "tailwindcss": "^4",
-    "typescript": "^5.4.0"
+    "tailwindcss": "^4"
   }
 }
